refactor(review/build): split tests to sibling tests_build.rs (#1266)
Inline test module in build.rs is moved into NEW sibling tests_build.rs.
Existing tests.rs (316 LOC) was not extended to avoid pushing it past 400.

- src/review/build.rs: 464 → 207 LOC (under 400-line rule)
- src/review/tests_build.rs: new (~257 LOC)
- src/review/mod.rs: register mod tests_build
- Widened to pub(super): failed_signal, now_unix_ms, review_summary, improvement_proposals
- super::* → super::build::* + use super::types::* + handoff/memory imports

Pure refactor: zero logic changes, all 11 tests pass.

Closes 9/70 of #1266 (this loop iteration), 18 PRs cumulative since #1285.
